Memory optimisations.
Reducing the size of an Action to 2 bytes seems reasonable:
if more moves are possible in a position then move grouping
should be used.

+++ include/casino.h
@@ -15,7 +15,7 @@
 extern "C" {
 #endif
 
-typedef uint64_t CAS_Action;
+typedef uint16_t CAS_Action;
 typedef void* CAS_DomainState;
 
 static const CAS_Action BAD_ACTION = ~(CAS_Action)0;
